# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Support Request Management API
  version: 1.0.0
  description: REST API used by internal agents to triage, assign and resolve customer
    support requests. Every mutation is audited and every resource is addressed by
    its public UUID.
paths:
...
  /api/v1/public/tickets/:
    post:
      operationId: public_tickets_create
      description: 'Unauthenticated intake. Throttled per IP and per reported email,
        and idempotent: send an `Idempotency-Key` header, or rely on the 60-second
        duplicate window. A replayed request answers `200` with the original ticket
        instead of creating a second one (ADR-15).'
      summary: Submit a support request
      parameters:
      - in: header
        name: Idempotency-Key
        schema:
          type: string
        description: Client-generated key; a repeat within 24h returns the same ticket.
      tags:
      - public
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PublicTicketCreateRequest'
        required: true
      security:
      - {}
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TicketPublic'
          description: ''
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TicketPublic'
          description: Duplicate submission replayed
        '429':
          description: '`throttled`'
  /api/v1/public/tickets/{public_id}/:
    get:
      operationId: public_tickets_retrieve
      description: Status and dates only. No agent identity, no comments, no history.
      summary: Track a support request
      parameters:
      - in: path
        name: public_id
        schema:
          type: string
          format: uuid
        required: true
      tags:
      - public
      security:
      - {}
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TicketPublic'
          description: ''
  /api/v1/tickets/:
    get:
      operationId: tickets_list
      description: |-
        The agent-facing ticket workflow.

        There is no ``DELETE``: a reported incident is a fact, and destroying it
        destroys the trail that made the system worth building (ADR-13).
      parameters:
      - in: query
        name: assignee
        schema:
          type: string
          format: uuid
      - in: query
        name: created_after
        schema:
          type: string
          format: date-time
      - in: query
        name: created_before
        schema:
          type: string
          format: date-time
      - in: query
        name: customer_email
        schema:
          type: string
      - name: ordering
        required: false
        in: query
        description: Which field to use when ordering the results.
        schema:
          type: string
      - name: page
        required: false
        in: query
        description: A page number within the paginated result set.
        schema:
          type: integer
      - name: page_size
        required: false
        in: query
        description: Number of results to return per page.
        schema:
          type: integer
      - in: query
        name: priority
        schema:
          type: array
          items:
            type: string
            enum:
            - HIGH
            - LOW
            - MEDIUM
            - URGENT
        description: |-
          * `LOW` - Low
          * `MEDIUM` - Medium
          * `HIGH` - High
          * `URGENT` - Urgent
        explode: true
        style: form
      - name: search
        required: false
        in: query
        description: A search term.
        schema:
          type: string
      - in: query
        name: status
        schema:
          type: array
          items:
            type: string
            enum:
            - CLOSED
            - IN_PROGRESS
            - OPEN
            - PENDING_CUSTOMER
            - RESOLVED
        description: |-
          * `OPEN` - Open
          * `IN_PROGRESS` - In progress
          * `PENDING_CUSTOMER` - Pending customer
          * `RESOLVED` - Resolved
          * `CLOSED` - Closed
        explode: true
        style: form
      - in: query
        name: unassigned
        schema:
          type: boolean
        description: Only unclaimed tickets
      tags:
      - tickets
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedTicketListList'
          description: ''
    post:
      operationId: tickets_create
      description: Accepts an optional `Idempotency-Key`; a replay answers `200` with
        the original ticket instead of creating a second one.
      summary: Create a ticket on behalf of a customer
      tags:
      - tickets
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AgentTicketCreateRequest'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/AgentTicketCreateRequest'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/AgentTicketCreateRequest'
        required: true
      security:
      - jwtAuth: []
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TicketDetail'
          description: ''
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TicketDetail'
          description: Duplicate request replayed
  /api/v1/tickets/{public_id}/:
    get:
      operationId: tickets_retrieve
      description: |-
        The agent-facing ticket workflow.

        There is no ``DELETE``: a reported incident is a fact, and destroying it
        destroys the trail that made the system worth building (ADR-13).
      parameters:
      - in: path
        name: public_id
        schema:
          type: string
          pattern: ^[0-9a-fA-F-]{36}$
        required: true
      tags:
      - tickets
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TicketDetail'
          description: ''
    patch:
      operationId: tickets_partial_update
      description: Only `subject` and `description`. State lives behind its own actions.
      summary: Edit the descriptive fields
      parameters:
      - in: header
        name: If-Match
        schema:
          type: integer
        description: The `version` the client last read. When present and stale the
          request is rejected with 409 `version_conflict` instead of silently overwriting
          somebody else's change.
      - in: path
        name: public_id
        schema:
          type: string
          pattern: ^[0-9a-fA-F-]{36}$
        required: true
      tags:
      - tickets
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PatchedTicketUpdateRequest'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/PatchedTicketUpdateRequest'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/PatchedTicketUpdateRequest'
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TicketDetail'
          description: ''
  /api/v1/tickets/{public_id}/assign/:
    post:
      operationId: tickets_assign_create
      description: '`assignee_id: null` releases the ticket; an `IN_PROGRESS` ticket
        then returns to `OPEN` (ADR-14).'
      summary: Assign, reassign or release
      parameters:
      - in: header
        name: If-Match
        schema:
          type: integer
        description: The `version` the client last read. When present and stale the
          request is rejected with 409 `version_conflict` instead of silently overwriting
          somebody else's change.
      - in: path
        name: public_id
        schema:
          type: string
          pattern: ^[0-9a-fA-F-]{36}$
        required: true
      tags:
      - tickets
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AssignRequest'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/AssignRequest'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/AssignRequest'
        required: true
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TicketDetail'
          description: ''
  /api/v1/tickets/{public_id}/assign-to-me/:
    post:
      operationId: tickets_assign_to_me_create
      description: Shortcut for the only assignment an agent makes all day.
      summary: Take the ticket
      parameters:
      - in: header
        name: If-Match
        schema:
          type: integer
        description: The `version` the client last read. When present and stale the
          request is rejected with 409 `version_conflict` instead of silently overwriting
          somebody else's change.
      - in: path
        name: public_id
        schema:
          type: string
          pattern: ^[0-9a-fA-F-]{36}$
        required: true
      tags:
      - tickets
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TicketDetail'
          description: ''
  /api/v1/tickets/{public_id}/comments/:
    get:
      operationId: tickets_comments_list
      description: |-
        The agent-facing ticket workflow.

        There is no ``DELETE``: a reported incident is a fact, and destroying it
        destroys the trail that made the system worth building (ADR-13).
      summary: List the comments
      parameters:
      - in: query
        name: assignee
        schema:
          type: string
          format: uuid
      - in: query
        name: created_after
        schema:
          type: string
          format: date-time
      - in: query
        name: created_before
        schema:
          type: string
          format: date-time
      - in: query
        name: customer_email
        schema:
          type: string
      - name: ordering
        required: false
        in: query
        description: Which field to use when ordering the results.
        schema:
          type: string
      - name: page
        required: false
        in: query
        description: A page number within the paginated result set.
        schema:
          type: integer
      - name: page_size
        required: false
        in: query
        description: Number of results to return per page.
        schema:
          type: integer
      - in: query
        name: priority
        schema:
          type: array
          items:
            type: string
            enum:
            - HIGH
            - LOW
            - MEDIUM
            - URGENT
        description: |-
          * `LOW` - Low
          * `MEDIUM` - Medium
          * `HIGH` - High
          * `URGENT` - Urgent
        explode: true
        style: form
      - in: path
        name: public_id
        schema:
          type: string
          pattern: ^[0-9a-fA-F-]{36}$
        required: true
      - name: search
        required: false
        in: query
        description: A search term.
        schema:
          type: string
      - in: query
        name: status
        schema:
          type: array
          items:
            type: string
            enum:
            - CLOSED
            - IN_PROGRESS
            - OPEN
            - PENDING_CUSTOMER
            - RESOLVED
        description: |-
          * `OPEN` - Open
          * `IN_PROGRESS` - In progress
          * `PENDING_CUSTOMER` - Pending customer
          * `RESOLVED` - Resolved
          * `CLOSED` - Closed
        explode: true
        style: form
      - in: query
        name: unassigned
        schema:
          type: boolean
        description: Only unclaimed tickets
      tags:
      - tickets
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedCommentList'
          description: ''
    post:
      operationId: tickets_comments_create
      description: |-
        The agent-facing ticket workflow.

        There is no ``DELETE``: a reported incident is a fact, and destroying it
        destroys the trail that made the system worth building (ADR-13).
      summary: Add a comment
      parameters:
      - in: path
        name: public_id
        schema:
          type: string
          pattern: ^[0-9a-fA-F-]{36}$
        required: true
      tags:
      - tickets
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CommentCreateRequest'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/CommentCreateRequest'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/CommentCreateRequest'
        required: true
      security:
      - jwtAuth: []
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Comment'
          description: ''
...
components:
  schemas:
...
    PaginatedTicketListList:
      type: object
      required:
      - count
      - results
      properties:
        count:
          type: integer
          example: 123
        next:
          type: string
          nullable: true
          format: uri
          example: http://api.example.org/accounts/?page=4
        previous:
          type: string
          nullable: true
          format: uri
          example: http://api.example.org/accounts/?page=2
        results:
          type: array
          items:
            $ref: '#/components/schemas/TicketList'
...
    PublicTicketCreateRequest:
      type: object
      description: |-
        What an unauthenticated customer is allowed to state.

        Note what cannot be set from here: status, assignee, and the effective
        priority. The customer reports a priority; triage decides one (ADR-09).
      properties:
        customer:
          $ref: '#/components/schemas/CustomerInputRequest'
        subject:
          type: string
          minLength: 3
          maxLength: 200
        description:
          type: string
          minLength: 10
          maxLength: 20000
        reported_priority:
          allOf:
          - $ref: '#/components/schemas/TicketPriorityEnum'
          default: MEDIUM
      required:
      - customer
      - description
      - subject
...
    TicketPublic:
      type: object
      description: |-
        What a customer may see about their own ticket.

        Subject, status and dates — no description of internal work, no agent
        identity, no history, no comments (ADR-03).
      properties:
        public_id:
          type: string
          format: uuid
          readOnly: true
        subject:
          type: string
          readOnly: true
        status:
          allOf:
          - $ref: '#/components/schemas/TicketStatusEnum'
          readOnly: true
        status_display:
          type: string
          readOnly: true
        created_at:
          type: string
          format: date-time
          readOnly: true
        updated_at:
          type: string
          format: date-time
          readOnly: true
        resolved_at:
          type: string
          format: date-time
          readOnly: true
          nullable: true
        closed_at:
          type: string
          format: date-time
          readOnly: true
          nullable: true
      required:
      - closed_at
      - created_at
      - public_id
      - resolved_at
      - status
      - status_display
      - subject
      - updated_at
...
  securitySchemes:
    jwtAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...